+++ cli/src/util/installer.ts
@@ -12,6 +12,7 @@ import path from "node:path";
 import * as fs from "./fs.js";
 import * as vi from "./versioninfo.js";
 import { extract } from "./tar.js";
+import { execFile } from "./cp.js";
 import { downloadFile, isOnline } from "./index.js";
 import { GitHubOwner, GitHubRepo, SDK } from "../custom/globals.js";
 
@@ -59,4 +60,35 @@ export async function installRuntime(version: string) {
   await fs.mkdir(installDir, { recursive: true });
   await extract(archivePath, installDir);
   await fs.rm(archivePath);
+}
+
+export async function installBuildTools(sdk: SDK, sdkVersion: string) {
+  switch (sdk) {
+    case SDK.Go:
+      await installGoBuildTools(sdkVersion);
+      break;
+  }
+}
+
+async function installGoBuildTools(sdkVersion: string) {
+  const sdkPath = vi.getSdkPath(SDK.Go, sdkVersion);
+
+  const ext = os.platform() === "win32" ? ".exe" : "";
+  const buildTool = path.join(sdkPath, "modus-go-build" + ext);
+  if (await fs.exists(buildTool)) {
+    return;
+  }
+
+  if (!(await isOnline())) {
+    throw new Error("Could not find the Modus Go build tool. Please try again when you are online.");
+  }
+
+  const module = `github.com/${GitHubOwner}/${GitHubRepo}/sdk/go/tools/modus-go-build@${sdkVersion}`;
+  await execFile("go", ["install", module], {
+    shell: true,
+    env: {
+      ...process.env,
+      GOBIN: sdkPath,
+    },
+  });
 }